cms/requirements.txt
Django==1.8.6
Pillow==3.0.0
wheel==0.24.0
django-storages
dj-database-url==0.3.0
dj_static
gunicorn==19.1.1
psycopg2==2.5.1
wsgiref==0.1.2
static==0.4
djangorestframework==3.3.1
python-instagram
requests
django-cors-headers